Added a key press listener
The editor listens for a specific key (curently Space) and performs fade
when it is pressed. This of course requires that the editor window needs
focus.

The click listener of the fade button was also moved to its own function
so it can also be used in the key listener.